Package processing moved to service, lockscreen fixes, continue to run on apps that use mic
Notice that Launchactivity is no longer used, but is kept incase we need
it later.
Also remember that the app will still not activate when inside OpenApps
(probably should change this...)
Finally having an issue where the lockscreen comes back after awhile. I
know a way around this, but will do it later...